15768 siemens_plc_duration: Fix crash if duration is smaller than a second
Previously, siemens_plc_duration check would crash if duration was smaller
than a second with an error:
C+:
ValueError (invalid literal for int() with base 10: '0.05000000074505806')
C-:

Now, the check doesn't crash anymore and it's possible to input thresholds
smaller than a second in the "Siemens PLC Duration" monitoring rule.

SUP-14553

+++ cmk/base/legacy_checks/siemens_plc.py
@@ -134,9 +134,9 @@ def check_siemens_plc_duration(item, params, info):
             2
         ] == item:
             if line[1].startswith("hours"):
-                seconds = int(line[-1]) * 3600
+                seconds = float(line[-1]) * 3600
             else:
-                seconds = int(line[-1])
+                seconds = float(line[-1])
 
             perfdata = [(line[1], seconds)]
